\name{RsegWord-package}
\alias{RsegWord-package}
\alias{RsegWord-package}
\title{A package to segment Chinese sentence}
\description{Use the imdict-chinese-analyzer (see \url{http://ictclas.org/news_ictclas_code.html}) to do words segmentation in R.}
\details{\tabular{ll}{ 
Package: \tab Rweibo\cr 
Type: \tab Package\cr 
Version: \tab 0.0-1\cr
Date: \tab 2011-08-03\cr 
License: \tab GPL\cr 
LazyLoad: \tab yes\cr 
}}
\alias{RsegWord-package}
\alias{RsegWord}
\docType{package}
\author{Jian li

Maintainer: Jian Li <rweibo@sina.com>}
\references{\url{http://ictclas.org/}}
\keyword{package}

