\name{rAnalysisCN-package}
\title{Segment Chinese Words.}
\description{Segment Chinese Words.
A package to segment Chinese words
description of the package.}
\details{\tabular{ll}{ Package: \tab rAnalysisCN\cr 
Type: \tab Package\cr 
Version: \tab 0.1\cr 
Date: \tab 2010-11-24\cr 
License: \tab GPL\cr 
LazyLoad: \tab yes\cr }
An overview of how to use the package, including the most
important ~~ ~~ functions ~~}
\alias{rAnalysisCN-package}
\alias{rAnalysisCN}
\docType{package}
\author{Jian Lee

Maintainer: Jian li <yourfault@somewhere.net>}
\seealso{~~ Optional links to other man pages, e.g. ~~ ~~
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~}
\references{~~ Literature or other references for background information ~~}
\keyword{package}
\examples{~~ simple examples of the most important functions ~~}
